Añadidos tests y reconfigurada serializacion dispatcher

+++ src/Editor/RetroAdventureCreator.Core/Models/GameComponentPointerModel.cs
@@ -2,4 +2,9 @@
 
 namespace RetroAdventureCreator.Core.Models;
 
-internal record GameComponentPointerModel(string Code, int RelativePointer) : IUniqueKey;
+/// <summary>
+/// Store game component (16 bits) pointer
+/// </summary>
+/// <param name="Code">Codigo del componente</param>
+/// <param name="RelativePointer">Relative pointer to component (16 bits)</param>
+internal record GameComponentPointerModel(string Code, short RelativePointer) : IUniqueKey;
